+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [1.0.3](https://github.com/akihisa1210/amazon-bibliographic-information-to-scrapbox/compare/v1.0.2...v1.0.3) (2020-08-24)
+
+
+### Bug Fixes
+
+* update packages for security ([e0a0124](https://github.com/akihisa1210/amazon-bibliographic-information-to-scrapbox/commit/e0a0124be91d44d97d56cfce03701cfefca9d3ca))
+* Update the id for detail bullet parts ([d5480c6](https://github.com/akihisa1210/amazon-bibliographic-information-to-scrapbox/commit/d5480c6c91660d09cff0b52509bf2d1dd133d374))
+
+
+
 ## [1.0.2](https://github.com/akihisa1210/amazon-bibliographic-information-to-scrapbox/compare/v1.0.1...v1.0.2) (2020-07-24)